OD-12357 admin, editor-api: add ON DELETE CASCADE for WebNodeType -> WebNode relation (problem with slave galera nodes, node can not delete WebNode)
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@55185 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/webdav/TemplateResourceFactory.groovy
@@ -21,7 +21,6 @@ import ish.oncourse.willow.editor.website.WebTemplateFunctions
 import org.apache.cayenne.ObjectContext
 import org.apache.cayenne.query.ObjectSelect
 import org.apache.commons.io.IOUtils
-import org.apache.commons.lang3.ArrayUtils
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.reflections.Reflections
@@ -228,8 +227,12 @@ class TemplateResourceFactory implements ResourceFactory {
         void delete() throws NotAuthorizedException, ConflictException, BadRequestException {
             ObjectContext context = cayenneService.newContext()
             WebSiteLayout layoutToChange = context.localObject(layout)
-            context.deleteObjects(layoutToChange)
-            context.commitChanges()
+            if (layoutToChange.webNodeTypes.empty) {
+                context.deleteObjects(layoutToChange)
+                context.commitChanges()
+            } else {
+                throw new BadRequestException(this, "Remove ${layoutToChange.webNodeTypes.size()} blocks from layout.".toString())
+            }
         }
     }
 }
